backend/requirements.txt
asgiref==3.10.0
asttokens==2.4.1
colorama==0.4.6
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
Django==5.2.8
djangorestframework==3.16.1
executing==2.0.1
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.3
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.11
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
#pywin32==306
#pywin32==306
pyzmq==25.1.2
setuptools==80.9.0
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
tzdata==2025.2
wcwidth==0.2.13
wheel==0.45.1
PyJWT
python-dotenv
supabase
pdfminer.six